tools/mockup_server_libvirt/requirements.txt
flask
libvirt-python
